Update package structure from `com.powsybl.network_area` to `com.powsybl.network.area`
