+++ balances-adjustment/src/main/java/com/powsybl/balances_adjustment/balance_computation/BalanceComputationArea.java
@@ -6,7 +6,7 @@
  */
 package com.powsybl.balances_adjustment.balance_computation;
 
-import com.powsybl.network_area.NetworkAreaFactory;
+import com.powsybl.network.area.NetworkAreaFactory;
 import com.powsybl.iidm.modification.scalable.Scalable;
 
 import java.util.Objects;
