+++ docs/balances_adjustment/index.md
@@ -127,7 +127,7 @@ import com.powsybl.balances_adjustment.balance_computation.*;
 import com.powsybl.computation.local.LocalComputationManager;
 import com.powsybl.iidm.network.Network;
 import com.powsybl.loadflow.LoadFlow;
-import com.powsybl.network_area.CountryAreaFactory;
+import com.powsybl.network.area.CountryAreaFactory;
 import com.powsybl.iidm.modification.scalable.GeneratorsScalable;
 
 import java.util.Arrays;
