+++ network-area/src/main/java/com/powsybl/network/area/InjectionBasedVoltageLevelsArea.java
@@ -5,7 +5,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  * SPDX-License-Identifier: MPL-2.0
  */
-package com.powsybl.network_area;
+package com.powsybl.network.area;
 
 import com.powsybl.iidm.network.Injection;
 import com.powsybl.iidm.network.Network;
