+++ network-area/src/main/java/com/powsybl/network/area/NetworkAreaFactory.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.network_area;
+package com.powsybl.network.area;
 
 import com.powsybl.iidm.network.Network;
