+++ network-area/src/test/java/com/powsybl/network/area/BorderBasedVoltageLevelsAreaTest.java
@@ -1,11 +1,10 @@
 /*
- * Copyright (c) 2025, RTE (http://www.rte-france.com)
+ * Copyright (c) 2019, RTE (http://www.rte-france.com)
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- * SPDX-License-Identifier: MPL-2.0
  */
-package com.powsybl.network_area;
+package com.powsybl.network.area;
 
 import com.powsybl.iidm.network.Network;
 import org.junit.jupiter.api.BeforeEach;
@@ -14,22 +13,22 @@ import org.junit.jupiter.api.Test;
 import java.util.ArrayList;
 import java.util.List;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 /**
- * @author Joris Mancini <joris.mancini_externe at rte-france.com>
+ * @author Ameni Walha {@literal <ameni.walha at rte-france.com>}
  */
-class InjectionBasedVoltageLevelsAreaTest {
+class BorderBasedVoltageLevelsAreaTest {
+
     private Network testNetwork;
     private VoltageLevelsAreaFactory voltageLevelsArea;
 
     @BeforeEach
     void setUp() {
-        testNetwork = Network.read(
-            "testCase.xiidm",
-            BorderBasedVoltageLevelsAreaTest.class.getResourceAsStream("/testCase.xiidm")
-        );
+        testNetwork = Network.read("testCase.xiidm", BorderBasedVoltageLevelsAreaTest.class.getResourceAsStream("/testCase.xiidm"));
+
         voltageLevelsArea = new VoltageLevelsAreaFactory("FFR1AA1", "DDE3AA1");
+
     }
 
     @Test
@@ -40,10 +39,7 @@ class InjectionBasedVoltageLevelsAreaTest {
         flows.add(testNetwork.getBranch("DDE1AA1  DDE3AA1  1").getTerminal2().getP());
         flows.add(testNetwork.getBranch("DDE2AA1  DDE3AA1  1").getTerminal2().getP());
 
-        assertEquals(
-            flows.stream().mapToDouble(f -> f).sum(),
-            voltageLevelsArea.create(testNetwork, true).getNetPosition(), 1e-3
-        );
+        assertEquals(flows.stream().mapToDouble(f -> f).sum(), voltageLevelsArea.create(testNetwork).getNetPosition(), 1e-3);
     }
 
     @Test
@@ -51,13 +47,13 @@ class InjectionBasedVoltageLevelsAreaTest {
         Network network = Network.read("testCaseSpecialDevices.xiidm", getClass().getResourceAsStream("/testCaseSpecialDevices.xiidm"));
 
         NetworkAreaFactory test3wtFactory = new VoltageLevelsAreaFactory("VOLTAGE_LEVEL_FR_225KV");
-        assertEquals(0, test3wtFactory.create(network, true).getNetPosition(), 1e-3);
+        assertEquals(0, test3wtFactory.create(network).getNetPosition(), 1e-3);
 
         NetworkAreaFactory testHvdcFactory = new VoltageLevelsAreaFactory("VOLTAGE_LEVEL_FR_225KV", "VOLTAGE_LEVEL_FR_400KV");
-        assertEquals(100, testHvdcFactory.create(network, true).getNetPosition(), 1e-3);
+        assertEquals(100, testHvdcFactory.create(network).getNetPosition(), 1e-3);
 
         NetworkAreaFactory test400kVFactory = new VoltageLevelsAreaFactory("VOLTAGE_LEVEL_FR_400KV", "VOLTAGE_LEVEL_BE_400KV");
-        assertEquals(100, test400kVFactory.create(network, true).getNetPosition(), 1e-3);
+        assertEquals(100, test400kVFactory.create(network).getNetPosition(), 1e-3);
 
     }
 }
